+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export `Fee`, `Fees`, `IndividualTxFees`, `MetaMetricsProps` and `FeatureFlags` types ([#000](https://github.com/MetaMask/smart-transactions-controller/pull/000))
+
 ## [19.0.0]
 
 ### Added